requirements.txt
apache-airflow-providers-http==4.10.0
pandas>=1.5.0,<2.1.0
requests>=2.25.0
pyarrow>=11.0.0
pandera>=0.13.0
pydantic>=2.0.0,<2.6.0
pytest>=8.0.0
psycopg2-binary